Create QR Code
Hi! I'm Samahar Sajjad!
This is my QR Code project that is most use and in most demanding in Every industry.
As you know about API (Application Programming Interface)....So, API is used in this Project...Here is my code of my Project with API Integration.....So, You can take help by using this code.
Languages that is used in this Project:
1.HTML
2.CSS
3.JAVASCRIPT
It's going to be so beneficial for you!
Thankyou!