Create Quiz Project
Hi! I'm Samahar Sajjad!
This is my Quiz Project.... In This Project, I used Javascript and create many elements...So, this project is like MCQS.....Here, is my code! You can take help by using my code.
Languages that is used in this Project:
1 HTML
2 CSS
3 JAVASCRIPT
It's going to be so beneficial for you!
Thankyou!!!